uv/tests: add several conflict marker tests
Some of these are regression tests or related regression tests
for #9289. Others are regression tests for bugs I found while working
on #9289.

Most of the outputs for these tests are wrong. For example, some tests
are asserting that multiple versions of the same package are installed.
But we add them now before our fixes so that we can see how they
change.