Improve file pinning comments
Not ideal still but it captures the reason why the code is the way it is.

Ideally we should remove file pins as it exists now entirely, we're doing a expensive clone (it's noticeable in profiles) of a type that we can still retrieve later only to discard most of them after the resolution is done; But that's out of scope for now.