Read publish username from URL
Setting the username in the URL was possible before, but google cloud artifacts would return a URL redirect for it to the URL without the username in it, and publish (POST) doesn't support URL redirect since the credentials get dropped on redirect.

Removing the username from the URL and setting it as regular username fixes this.

See #9778