Remove unnecessary build backend from tests (#10868)
## Summary

These tests don't need a build backend. If we omit it, the project is
treated as virtual, and we avoid building and installing it.

The only changes in the snapshots should be a decrement in resolve or
install count, since we're often now omitting the project itself.

I left the build backend for anything borderline, including workspace
members within tests.